HUD adjustments/bug fixes
Set default text so that you can see the skeleton HUD without values in editor. Removed GameManager object from all levels except MainMenu because it was causing problems with the score/lives systems as there were multiple instances of the same GameManager added each time you died or moved to the next level. Renamed a few things in the HUD scripts as well as in the editor for clarity. Fixed bug where score would not reset after losing or returning to the main menu from the pause screen. Fixed pause menu text on the Resume button to match the main menu button like it should have. Moved the Sounds folder back outside of the Scenes folder. Keep in mind, now that GameManager is only created in the main menu, when testing you must start on the main menu and go from there.